# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/condominios/{id}": {
      "get": {
        "operationId": "CondominiosController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "condominios"
        ]
      },
      "put": {
        "operationId": "CondominiosController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCondominioDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "condominios"
        ]
      }
    },
...
    "/unidades": {
      "post": {
        "description": "Crea una nueva unidad en el condominio detectado del subdominio. Requiere rol SUPERADMIN o ADMIN.",
        "operationId": "UnidadesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUnidadDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Unidad creada exitosamente"
          },
          "400": {
            "description": "Datos inválidos o subdominio no detectado"
          },
          "403": {
            "description": "No autorizado - Se requiere rol SUPERADMIN o ADMIN"
          }
        },
        "security": [
          {
            "better-auth.session_token": []
          },
          {
            "JWT-auth": []
          }
        ],
        "summary": "Crear una nueva unidad",
        "tags": [
          "unidades"
        ]
      },
      "get": {
        "description": "Retorna todas las unidades del condominio detectado del subdominio",
        "operationId": "UnidadesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lista de unidades obtenida exitosamente"
          }
        },
        "security": [
          {
            "better-auth.session_token": []
          },
          {
            "JWT-auth": []
          }
        ],
        "summary": "Obtener todas las unidades",
        "tags": [
          "unidades"
        ]
      }
    },
    "/unidades/with-residentes": {
      "get": {
        "description": "Retorna todas las unidades del condominio incluyendo información de residentes asociados",
        "operationId": "UnidadesController_findAllWithResidentes",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lista de unidades con residentes obtenida exitosamente"
          }
        },
        "security": [
          {
            "better-auth.session_token": []
          },
          {
            "JWT-auth": []
          }
        ],
        "summary": "Obtener todas las unidades con residentes",
        "tags": [
          "unidades"
        ]
      }
    },
    "/unidades/{unidadId}": {
      "get": {
        "description": "Retorna los detalles de una unidad específica",
        "operationId": "UnidadesController_findOne",
        "parameters": [
          {
            "name": "unidadId",
            "required": true,
            "in": "path",
            "description": "ID de la unidad",
            "schema": {
              "example": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Unidad obtenida exitosamente"
          },
          "404": {
            "description": "Unidad no encontrada"
          }
        },
        "security": [
          {
            "better-auth.session_token": []
          },
          {
            "JWT-auth": []
          }
        ],
        "summary": "Obtener una unidad por ID",
        "tags": [
          "unidades"
        ]
      },
      "put": {
        "operationId": "UnidadesController_update",
        "parameters": [
          {
            "name": "unidadId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUnidadDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "unidades"
        ]
      },
      "patch": {
        "operationId": "UnidadesController_patch",
        "parameters": [
          {
            "name": "unidadId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUnidadDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "unidades"
        ]
      },
      "delete": {
        "operationId": "UnidadesController_delete",
        "parameters": [
          {
            "name": "unidadId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "unidades"
        ]
      }
    },
    "/unidades/bulk-upload": {
      "post": {
        "description": "Crea múltiples unidades en el condominio mediante carga masiva",
        "operationId": "UnidadesController_bulkUpload",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkUploadUnidadesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Unidades cargadas exitosamente"
          },
          "400": {
            "description": "Datos inválidos"
          }
        },
        "security": [
          {
            "better-auth.session_token": []
          },
          {
            "JWT-auth": []
          }
        ],
        "summary": "Cargar múltiples unidades",
        "tags": [
          "unidades"
        ]
      }
    }
  },
  "info": {
    "title": "API Vekino",
    "description": "API para gestión de condominios",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor de desarrollo"
    },
    {
      "url": "https://api.vekino.com",
      "description": "Servidor de producción"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Token de autenticación JWT",
        "in": "header"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token",
        "description": "Cookie de sesión de Better Auth"
      }
    },
    "schemas": {
...
      "UpdateCondominioDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateUnidadDto": {
        "type": "object",
        "properties": {
          "identificador": {
            "type": "string",
            "description": "Identificador único de la unidad",
            "example": "Apto 801"
          },
          "tipo": {
            "type": "string",
            "description": "Tipo de unidad",
            "enum": [
              "APARTAMENTO",
              "CASA",
              "LOCAL_COMERCIAL"
            ],
            "example": "APARTAMENTO"
          },
          "area": {
            "type": "number",
            "description": "Área de la unidad en metros cuadrados",
            "example": 75.5,
            "minimum": 0
          },
          "coeficienteCopropiedad": {
            "type": "number",
            "description": "Coeficiente de copropiedad en porcentaje",
            "example": 15.5,
            "minimum": 0,
            "maximum": 100
          },
          "valorCuotaAdministracion": {
            "type": "number",
            "description": "Valor de la cuota de administración",
            "example": 250000,
            "minimum": 0
          },
          "estado": {
            "type": "string",
            "description": "Estado actual de la unidad",
            "enum": [
              "OCUPADA",
              "VACIA",
              "EN_MANTENIMIENTO"
            ],
            "example": "OCUPADA"
          }
        },
        "required": [
          "identificador",
          "tipo"
        ]
      },
      "UpdateUnidadDto": {
        "type": "object",
        "properties": {}
      },
      "BulkUploadUnidadesDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}